refactor: 디자인 변경 사항 반영 및 코드 스타일 수정(#161)
